ingestion/demo_data_producer.py
import logging
from kafka import KafkaProducer
import json
import csv
from avro import schema, datafile, io

# Initialize Kafka producer
producer = KafkaProducer(bootstrap_servers='your_kafka_brokers')

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


def produce_message(topic, message):
    try:
        producer.send(topic, message.encode('utf-8'))
        producer.flush()
        logger.info(f"Message successfully produced to {topic}")
    except Exception as e:
        logger.error(f"Error producing message to {topic}: {e}")


# Ingest Ad Impressions (JSON)
ad_impressions = {"ad_creative_id": 1, "user_id": 123, "timestamp": "2024-03-07", "website": "example.com"}
produce_message('ad_impressions_topic', json.dumps(ad_impressions))

# Ingest Clicks/Conversions (CSV)
click_conversion_data = "timestamp,user_id,ad_campaign_id,conversion_type\n2024-03-07,123,456,signup"
produce_message('click_conversion_topic', click_conversion_data)

# Ingest Bid Requests (Avro)
schema_definition = schema.Parse(open('bid_request.avsc').read())
writer = datafile.DataFileWriter(open('bid_requests.avro', 'wb'), io.DatumWriter(), schema_definition)
bid_request_data = {"user_info": "some_info", "auction_details": "details", "ad_targeting_criteria": "criteria"}
try:
    writer.append(bid_request_data)
    logger.info("Bid request data successfully appended")
except Exception as e:
    logger.error(f"Error appending bid request data: {e}")
finally:
    writer.close()
